+++ package.json
@@ -7,6 +7,7 @@
     "start": "cross-env NODE_ENV=development node index.js",
     "build": "cross-env NODE_ENV=production node index.js",
     "publish": "neocities push ./build",
+    "start-server": "node ./server/index.js",
     "get-schema": "node schema/schemaQuery.js",
     "dev-static": "nodemon --config ./nodemon.static.config.json",
     "copy-static": "node ./utils/copy-static.js",
